Override DEFAULT_PERSISTENT_VOLUME_CLAIM value
The release operator will create PipelineRuns with a workspace.
This variable contains the claim to use for that workspace.
Another variable, DEFAULT_WORKSPACE_NAME controls the name of the
default workspace. This variable is populated by the controller
if not overridden here.

In case of DEFAULT_PERSISTENT_VOLUME_CLAIM not being overridden,
the workspace will be not created.

Signed-off-by: David Moreno García <damoreno@redhat.com>